Revision 2: Updated syntax, clarification, files
