+++ backend/src/ml_space_lambda/hpo_job/lambda_functions.py
@@ -24,12 +24,7 @@ from ml_space_lambda.data_access_objects.project_user import ProjectUserDAO
 from ml_space_lambda.data_access_objects.resource_metadata import ResourceMetadataDAO
 from ml_space_lambda.enums import ResourceType
 from ml_space_lambda.utils.common_functions import api_wrapper, generate_tags, query_resource_metadata, retry_config
-
-# The sagemaker SDK isn't directly compatible with Lambda
-# https://github.com/aws/sagemaker-python-sdk/issues/1200
-# To get around this we just grabbed the single file of the SDK that we need in order to generate
-# image URIs
-from ml_space_lambda.utils.image_uris import check_algorithm_specifications_for_builtin
+from ml_space_lambda.utils.image_uri_utils import delete_metric_definition_for_builtin_algorithms
 from ml_space_lambda.utils.mlspace_config import get_environment_variables, pull_config_from_s3
 
 logger = logging.getLogger(__name__)
@@ -57,7 +52,7 @@ def _normalize_job_definition(definition, iam_role, param_file):
         del definition["subnetIds"]
     else:
         definition["VpcConfig"]["Subnets"] = random.sample(param_file["pSMSSubnetIds"].split(","), 1)
-    check_algorithm_specifications_for_builtin(definition["AlgorithmSpecification"])
+    delete_metric_definition_for_builtin_algorithms(definition["AlgorithmSpecification"])
     return definition
